Subscribe section corrected
