+++ src/pages/Blog/Blog.tsx
@@ -1,13 +1,10 @@
+import Section from 'ui/Section/Section'
+import { BlogHero, BlogWrapper } from './Blog.styles'
+import Subscribe from './Subscribe/Subscribe'
 import { useState } from 'react'
 
-import Section from 'ui/Section/Section'
-import { BlogHero, BlogWrapper, SubscribeWrapper } from './Blog.styles'
-import subscribeImg from './../../assets/images/imgBlog/subscribe.svg'
-import Button from 'ui/Button/Button'
-import Input from 'ui/input/Input'
-
 const Blog: React.FC = () => {
-  const [value, setValue] = useState('')
+  const [text, setText] = useState('')
   return (
     <BlogWrapper size="xl" isfullwidth>
       <Section size="sm" type="margin">
@@ -20,37 +17,7 @@ const Blog: React.FC = () => {
         {/*questionable if the search will be performed within the component pagination */}
       </div>
       <div>Pagination component</div>
-      <Section type="margin" size="sm">
-        <SubscribeWrapper size="lg" isfullwidth>
-          <div className="subscribe-content">
-            <div>
-              <h2 className="subscribe-title">Suscríbete</h2>
-              <p className="subscribe-description">
-                ¿ Deseas estar mas pendiente de más artículos ?
-              </p>
-            </div>
-            <div className="subscribe-form">
-              <Input
-                type="email"
-                id="email"
-                value={value}
-                onChange={(e) => setValue(e.target.value)}
-                placeholder="Tu e-mail"
-                className="subscribe-input"
-              />
-              <Button
-                type="primary"
-                size="lg"
-                className="subscribe-button"
-                onClick={() => console.log(value)}
-              >
-                Suscríbete
-              </Button>
-            </div>
-          </div>
-          <img src={subscribeImg} className="subscribe-img" />
-        </SubscribeWrapper>
-      </Section>
+      <Subscribe text={text} setText={setText} />
     </BlogWrapper>
   )
 }
